Move functionalities from routes to controller
